main class code  change

+++ GITProj01/src/main/java/com/test/App.java
@@ -8,7 +8,7 @@ import com.nt.service.Arithmetic;
 public class App {
     public static void main(String[] args) {
         Arithmetic  ar=new Arithmetic();
-        int result=ar.sum(10, 20);
+        long result=ar.sum(10, 20);
         System.out.println("result::"+result);
     }
 }
